Refactored coverage merging pipeline.
